clean up images
