+++ frontend/src/components/images/AwardImage.tsx
@@ -1,0 +1,16 @@
+import { Image } from "./Image";
+
+type AwardImageProps = {
+  id: string | undefined;
+  size: AwardImageSize;
+};
+
+const sizeMap: Record<AwardImageSize, number> = {
+  s: 32,
+};
+
+type AwardImageSize = "s";
+
+export const AwardImage = ({ id, size }: AwardImageProps) => {
+  return <Image id={id} size={sizeMap[size]} />;
+};
